chapter-8/requirements.txt
redis==4.5.4
python-dotenv==0.21.1